requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==3.2.0
cffi==1.17.1
click==8.2.1
colorama==0.4.6
cryptography==45.0.2
dnspython==2.7.0
dotenv==0.9.9
ecdsa==0.19.1
email_validator==2.2.0
fastapi>=0.68.0
h11==0.16.0
idna==3.10
passlib>=1.7.4
pyasn1==0.4.8
pycparser==2.22
pydantic>=1.9.0
pydantic_core==2.33.2
python-dotenv==1.1.0
python-jose>=3.3.0
python-multipart>=0.0.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqlalchemy>=1.4.23
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.13.2
uvicorn>=0.15.0
streamlit>=1.28.0
requests>=2.31.0
pandas>=2.0.0